cleanups(): - added /renamed (to) index.js files - added handling of sub formats (binary vs ascii stl etc) - various cleanups
