+++ packages/deserializers/svg-deserializer/index.js
