+++ packages/serializers/json-serializer/package.json
@@ -33,6 +33,7 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@jscad/csg": "latest"
   },
   "devDependencies": {
   }
