+++ packages/serializers/stl-serializer/CSGToStlb.js
@@ -1,7 +1,6 @@
-const mimeType = 'application/sla'
 
 // see http://en.wikipedia.org/wiki/STL_%28file_format%29#Binary_STL
-function write (CSG) {
+function serialize (CSG) {
   // first check if the host is little-endian:
   var buffer = new ArrayBuffer(4)
   var int32buffer = new Int32Array(buffer, 0, 1)
@@ -57,13 +56,12 @@ function write (CSG) {
       byteoffset += 50
     }
   })
-  return [headerarray.buffer, ar1.buffer, allTrianglesBuffer]//'blobable array'
-  /*return new Blob([headerarray.buffer, ar1.buffer, allTrianglesBuffer], {
+  return [headerarray.buffer, ar1.buffer, allTrianglesBuffer]// 'blobable array'
+  /* return new Blob([headerarray.buffer, ar1.buffer, allTrianglesBuffer], {
     type: mimeType
-  })*/
+  }) */
 }
 
 module.exports = {
-  write,
-  mimeType
+  serialize
 }
