+++ packages/serializers/svg-serializer/index.js
@@ -3,7 +3,7 @@ const {CSG} = require('@jscad/csg')
 
 const mimeType = 'image/svg+xml'
 
-function write (cagObject) {
+function serialize (cagObject) {
   var decimals = 1000
 
   // mirror the CAG about the X axis in order to generate paths into the POSITIVE direction
@@ -49,6 +49,6 @@ function PathsToSvg (paths, bounds) {
 }
 
 module.exports = {
-  write,
+  serialize,
   mimeType
 }
